--- src/main/java/com/prgrms/amabnb/oauth/dto/AccessTokenResponse.java
@@ -1,4 +1,0 @@
-package com.prgrms.amabnb.oauth.dto;
-
-public record AccessTokenResponse(String accessToken) {
-}
