--- src/main/java/com/prgrms/amabnb/oauth/dto/TokenResponse.java
@@ -1,4 +1,0 @@
-package com.prgrms.amabnb.oauth.dto;
-
-public record TokenResponse(String accessToken, String refreshToken) {
-}
