+++ src/main/java/com/prgrms/amabnb/security/handler/OAuthAuthenticationSuccessHandler.java
@@ -12,8 +12,8 @@ import org.springframework.security.web.authentication.AuthenticationSuccessHand
 import org.springframework.stereotype.Component;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.prgrms.amabnb.oauth.service.OAuthProvider;
-import com.prgrms.amabnb.oauth.service.OAuthService;
+import com.prgrms.amabnb.oauth.OAuthProvider;
+import com.prgrms.amabnb.oauth.OAuthService;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
