+++ src/main/java/com/prgrms/amabnb/token/dto/AccessTokenResponse.java
@@ -1,0 +1,4 @@
+package com.prgrms.amabnb.token.dto;
+
+public record AccessTokenResponse(String accessToken) {
+}
