+++ src/main/java/com/prgrms/amabnb/token/dto/TokenResponse.java
@@ -1,0 +1,4 @@
+package com.prgrms.amabnb.token.dto;
+
+public record TokenResponse(String accessToken, String refreshToken) {
+}
